+++ packages/python-client/src/redenv/__init__.py
@@ -3,5 +3,5 @@ from .errors import RedenvError
 from .secrets import Secrets
 from .sync import Redenv as RedenvSync
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 __all__ = ["Redenv", "RedenvSync", "RedenvError", "Secrets"]